refactor(diff): split hljs engine patch out of gap-isolation
The highlight.js singleton patch moves to a new `hljs-gap-isolation.ts`,
leaving `gap-isolation.ts` with type-only imports from
`@git-diff-view/core`. That module is reachable from the highlight worker
(highlight-worker -> shiki-highlighter -> here), so keeping it free of
engine code keeps the worker chunk from pulling in the highlighter (and
React, via the react entry). `DiffSurface` now imports
`installHljsGapIsolation` from its new home.

`isHoley` and `mergeSegments` are exported for that one consumer, and
`mergeSegments` returns null instead of re-tokenizing when the merged AST
doesn't flatten back to its input — each caller falls back to the
whole-buffer result it already has, so `highlightAuto` no longer pays for
a third tokenize pass. The merged root also inherits the first segment's
`data`, since every segment is tokenized as the same language.
